Add admin authentication functionality
This commit introduces a crucial function - the ability for admins to authenticate. It includes the creation of a 'comparePassword' utility on 'util.js', incorporation of the 'bcryptjs' dependency, and modifications to several files such as 'authMiddleware.js', 'admin.js', and 'auth.js' to enable the feature. The jwtAdminMiddleware was also edited to ensure smooth functioning of the admin role.